Fixed an additional unmount edge case. Moved hook into temporary package.
