+++ packages/react-hooks/index.js
@@ -9,4 +9,4 @@
 
 'use strict';
 
-export * from './src/createSubscription';
+export * from './src/useSubscription';
